requirements.txt
requests>=2.31.0
pandas>=2.0.0
openai>=1.97.0
python-dotenv>=1.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
numpy>=1.24.0
